requirements.txt
# Conversor HB → Painel CCM
# Dependências Python

pandas>=2.0.0
openpyxl>=3.1.0
PyYAML>=6.0
